Improve verifier ut and rename test configuration
