+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/AadRestTemplateCreatorTest.java
@@ -31,7 +31,7 @@ import java.util.List;
 
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.AadRestTemplateCreator.createOAuth2AccessTokenResponseClientRestTemplate;
 import static com.azure.spring.cloud.autoconfigure.aad.implementation.AadRestTemplateCreator.createOAuth2ErrorResponseHandledRestTemplate;
-import static com.azure.spring.cloud.autoconfigure.aad.implementation.RestTemplateProxyCustomizerConfiguration.FACTORY;
+import static com.azure.spring.cloud.autoconfigure.aad.implementation.RestTemplateProxyCustomizerTestConfiguration.FACTORY;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.isA;
@@ -106,7 +106,7 @@ class AadRestTemplateCreatorTest {
         new ApplicationContextRunner()
                 .withUserConfiguration(
                         RestTemplateAutoConfiguration.class,
-                        RestTemplateProxyCustomizerConfiguration.class)
+                        RestTemplateProxyCustomizerTestConfiguration.class)
                 .run((context) -> {
                     RestTemplate restTemplate = context.getBean(RestTemplateBuilder.class).build();
                     assertSame(restTemplate.getRequestFactory(), FACTORY);
