+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/aad/implementation/oauth2/AadOAuth2ClientConfigurationTests.java
@@ -5,7 +5,7 @@ package com.azure.spring.cloud.autoconfigure.aad.implementation.oauth2;
 
 import com.azure.spring.cloud.autoconfigure.aad.AadClientRegistrationRepository;
 import com.azure.spring.cloud.autoconfigure.aad.configuration.AadOAuth2ClientConfiguration;
-import com.azure.spring.cloud.autoconfigure.aad.implementation.RestTemplateProxyCustomizerConfiguration;
+import com.azure.spring.cloud.autoconfigure.aad.implementation.RestTemplateProxyCustomizerTestConfiguration;
 import com.azure.spring.cloud.autoconfigure.aad.implementation.TestJwks;
 import com.azure.spring.cloud.autoconfigure.aad.implementation.webapi.AadJwtBearerGrantRequestEntityConverter;
 import com.azure.spring.cloud.autoconfigure.aad.properties.AadAuthenticationProperties;
@@ -228,7 +228,7 @@ class AadOAuth2ClientConfigurationTests {
     @Test
     void restTemplateWellConfiguredWhenNotUsingPrivateKeyJwtMethod() {
         webApplicationContextRunner()
-                .withUserConfiguration(AadOAuth2ClientConfiguration.class, RestTemplateProxyCustomizerConfiguration.class)
+                .withUserConfiguration(AadOAuth2ClientConfiguration.class, RestTemplateProxyCustomizerTestConfiguration.class)
                 .run(context -> {
                     assertThat(context).doesNotHaveBean(OAuth2ClientAuthenticationJwkResolver.class);
                     assertRestTemplateWellConfigured(context);
@@ -246,7 +246,7 @@ class AadOAuth2ClientConfigurationTests {
                         "spring.cloud.azure.active-directory.authorization-clients.graph.scopes=https://graph.microsoft.com/User.Read",
                         "spring.cloud.azure.active-directory.authorization-clients.graph.scopes=api://52261059-e515-488e-84fd-a09a3f372814/File.Read"
                 )
-                .withUserConfiguration(AadOAuth2ClientConfiguration.class, RestTemplateProxyCustomizerConfiguration.class)
+                .withUserConfiguration(AadOAuth2ClientConfiguration.class, RestTemplateProxyCustomizerTestConfiguration.class)
                 .run(context -> {
                     assertThat(context).hasSingleBean(OAuth2ClientAuthenticationJwkResolver.class);
                     assertRestTemplateWellConfigured(context);
